@@ -9,7 +9,7 @@ void main() => runApp(const MyApp());
 /// An empty app.
 class MyApp extends StatelessWidget {
   /// Creates an empty app.
-  const MyApp({Key key}) : super(key: key);
+  const MyApp({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
