@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 //
-// Autogenerated from Pigeon (v3.0.4), do not edit directly.
+// Autogenerated from Pigeon (v4.0.2), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 
 package dev.flutter.aaclarke.pigeon;
@@ -26,9 +26,9 @@ import java.util.Map;
 public class MessagePigeon {
 
   public enum MessageRequestState {
-    pending(0),
-    success(1),
-    failure(2);
+    PENDING(0),
+    SUCCESS(1),
+    FAILURE(2);
 
     private int index;
 
