@@ -2,10 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 //
-// Autogenerated from Pigeon (v3.2.0), do not edit directly.
+// Autogenerated from Pigeon (v4.0.2), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 // ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, prefer_null_aware_operators, omit_local_variable_types, unused_shown_name, unnecessary_import
-// @dart = 2.12
 import 'dart:async';
 import 'dart:typed_data' show Float64List, Int32List, Int64List, Uint8List;
 
@@ -66,7 +65,7 @@ class Everything {
     return Everything(
       aBool: pigeonMap['aBool'] as bool?,
       anInt: pigeonMap['anInt'] as int?,
-      aDouble: (pigeonMap['aDouble'] as num?)?.toDouble(),
+      aDouble: pigeonMap['aDouble'] as double?,
       aString: pigeonMap['aString'] as String?,
       aByteArray: pigeonMap['aByteArray'] as Uint8List?,
       a4ByteArray: pigeonMap['a4ByteArray'] as Int32List?,
